+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Infra: detach EBS volumes `awless detach volume`
 - Config: enable/disable the syncing of Route53 service `awless config set aws.dns.sync`
 - All listing with default format are now Markdown table compatible. 
+- Better display of `awless show`. Added `--siblings` flag to display exhaustively all siblings
 
 ## 0.0.18 [2017-03-13]